--- a/.env.example
+++ b/.env.example
@@ -4,6 +4,7 @@ APP_ID=
 PUBLIC_KEY=
 BOT_TOKEN=
 FAUNA_SECRET=
+IMAGE_PROXY_URL=
 
 # Recommended for local development