--- a/src/interactions.ts
+++ b/src/interactions.ts
@@ -1,3 +1,5 @@
+const ts = Date.now();
+
 import * as discord from './discord.ts';
 
 import search, { idPrefix } from './search.ts';
@@ -26,6 +28,8 @@ import { NonFetalError, NoPermissionError } from './errors.ts';
 
 export const handler = async (r: Request) => {
   const { origin } = new URL(r.url);
+
+  console.log(`request ready to handle - ${Date.now() - ts}ms`);
 
   // redirect to /demo on browsers
   if (
@@ -1108,7 +1112,7 @@ if (import.meta.main) {
   utils.serve({
     '/': handler,
     '/demo': demo,
-    '/external/*': utils.proxy,
+    // '/external/*': utils.proxy,
     '/webhooks/topgg': webhooks.topgg,
     '/invite': () =>
       Response.redirect(