--- a/src/utils.ts
+++ b/src/utils.ts
@@ -1,8 +1,6 @@
 // @deno-types="https://raw.githubusercontent.com/greggman/unzipit/v1.4.3/dist/unzipit.module.d.ts"
 import { unzip } from 'https://raw.githubusercontent.com/greggman/unzipit/v1.4.3/dist/unzipit.module.js';
 
-// import * as imagescript from 'https://deno.land/x/imagescript@1.2.15/mod.ts';
-
 import nacl from 'https://esm.sh/tweetnacl@1.0.3';
 
 import {
@@ -24,8 +22,6 @@ import {
 
 import { distance as _distance } from 'https://raw.githubusercontent.com/ka-weihe/fastest-levenshtein/1.0.15/mod.ts';
 
-// import { inMemoryCache } from 'https://deno.land/x/httpcache@0.1.2/in_memory.ts';
-
 import { RECHARGE_MINS } from '../models/get_user_inventory.ts';
 
 export enum ImageSize {
@@ -34,10 +30,6 @@ export enum ImageSize {
   Medium = 'medium', // 230x325
   Large = 'large', // 450x635,
 }
-
-// const TEN_MIB = 1024 * 1024 * 10;
-
-// const globalCache = inMemoryCache(20);
 
 function randint(min: number, max: number): number {
   return Math.floor(Math.random() * (max - min + 1) + min);
@@ -232,109 +224,6 @@ function verifySignature(
 
   return { valid, body };
 }
-
-// async function proxy(r: Request): Promise<Response> {
-//   const { pathname, searchParams } = new URL(r.url);
-
-//   try {
-//     let cached = true;
-
-//     const url = new URL(
-//       decodeURIComponent(pathname.substring('/external/'.length)),
-//     );
-
-//     const response = await globalCache.match(url as unknown as Request) ??
-//       (cached = false, await fetch(url, { signal: AbortSignal.timeout(3000) }));
-
-//     const type = response?.headers.get('content-type');
-
-//     const size = searchParams.get('size') as ImageSize || ImageSize.Large;
-
-//     if (Number(response.headers.get('content-length')) > TEN_MIB) {
-//       throw new Error();
-//     }
-
-//     // if (type === 'image/gif' && !url.pathname.endsWith('.gif')) {
-//     if (type === 'image/gif') {
-//       throw new Error();
-//     }
-
-//     if (response?.status !== 200 || !type?.startsWith('image/')) {
-//       throw new Error();
-//     }
-
-//     if (searchParams.has('blur')) {
-//       throw new Error();
-//     }
-
-//     const original = await response.arrayBuffer();
-
-//     if (!cached) {
-//       await globalCache.put(url as unknown as Request, new Response(original));
-//     }
-
-//     const transformed = await imagescript.decode(original);
-
-//     let proxy: Response | undefined = undefined;
-
-//     if (transformed instanceof imagescript.Image) {
-//       switch (size) {
-//         case ImageSize.Preview:
-//           transformed.cover(32, 32);
-//           break;
-//         case ImageSize.Thumbnail:
-//           transformed.cover(110, 155);
-//           break;
-//         case ImageSize.Medium:
-//           transformed.cover(230, 325);
-//           break;
-//         default:
-//           transformed.cover(450, 635);
-//           break;
-//       }
-
-//       if (type === 'image/png') {
-//         const t = await transformed.encode(2);
-
-//         proxy = new Response(t);
-
-//         proxy.headers.set('content-type', 'image/png');
-//         proxy.headers.set('content-length', `${t.byteLength}`);
-//       } else {
-//         const t = await transformed.encodeJPEG(70);
-
-//         proxy = new Response(t);
-
-//         proxy.headers.set('content-type', 'image/jpeg');
-//         proxy.headers.set('content-length', `${t.byteLength}`);
-//       }
-//     }
-
-//     if (!proxy) {
-//       throw new Error();
-//     }
-
-//     proxy.headers.set('cache-control', 'public, max-age=604800');
-
-//     return proxy;
-//   } catch {
-//     let fileUrl = new URL('../assets/medium.png', import.meta.url);
-
-//     if (r.url?.includes('?size=thumbnail')) {
-//       fileUrl = new URL('../assets/thumbnail.png', import.meta.url);
-//     }
-
-//     const body = await Deno.readFile(fileUrl);
-
-//     const response = new Response(body);
-
-//     response.headers.set('content-type', 'image/png');
-//     response.headers.set('cache-control', 'public, max-age=604800');
-//     response.headers.set('content-length', `${body.byteLength}`);
-
-//     return response;
-//   }
-// }
 
 async function readJson<T>(filePath: string): Promise<T> {
   try {