This PR addresses rare cases where Fable fails to respond in under 3 seconds causing discord to fail the interaction.

Deno Deploy colds start times range from 300ms to 1.2s, Fable used to add more to those due to the initialization step of a few dependencies, with this PR all requests should be responded to within an average of 600ms. Of course, there will always still be cases where this isn't true but that should be very rare and dependent on the status of Deno.

- replace the builtin image proxy that used "imagescript" with a standalone service [images-proxy](https://github.com/fable-community/images-proxy)
imagescript slowed the cold starts by about 50% due to how the denoified version fetches external wasm modules
deno deploy doesn't support dynamic imports so imagescript had to be imported for each request even if it wasn't required

- switched to a non-wasm ed25519 library
the effect of this is doubtful but it was removed in case initializing the module caused any slowdowns

- updated the fauna library to the esm.sh version of the official library